functions/11-governance/requirements.txt
functions-framework
google-cloud-pubsub
google-cloud-aiplatform
google-cloud-firestore
